requirements.txt
aiogram
moviepy
python-decouple
